+++ src/index.ts
@@ -9,6 +9,7 @@ export { toSchemaOrg } from './toSchemaOrg';
 export { extractSchemaOrgRecipeFromHTML } from './scraper/browser';
 export { SOUSTACK_SPEC_VERSION } from './specVersion';
 export { SOUSTACK_VERSION } from './version';
+export * as bundle from './bundle';
 export {
   convertLineItemToMetric,
   type ConvertMode,